Reduce dictionary size while  still accounting for length
Change the network input to take in a musical element (note, chord, rest,
metronome mark) and the length of that musical element. Change the network
output to output a musical element and its length. This reduces the dictionary
size since we would only have 480 possible musical elements and 25 possible
different lengths.